chore : rename schedules.ts to schedulte.type.ts #164
