+++ type/schedules/schedules.type.ts
